# packages/mtg-core/src/mtg_core/api/schema/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MTG Core API",
    "description": "Magic: The Gathering card database and analysis API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/recommendations/cards": {
      "post": {
        "tags": [
          "recommendations"
        ],
        "summary": "Suggest Cards For Deck",
        "description": "Suggest cards to add to a deck based on themes and synergies.",
        "operationId": "suggest_cards_for_deck_recommendations_cards_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuggestCardsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuggestCardsResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/recommendations/commanders": {
      "post": {
        "tags": [
          "recommendations"
        ],
        "summary": "Find Commanders",
        "description": "Find commanders that work well with a collection.",
        "operationId": "find_commanders_recommendations_commanders_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FindCommandersRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "type": "array",
                  "title": "Response Find Commanders Recommendations Commanders Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/recommendations/decks": {
      "post": {
        "tags": [
          "recommendations"
        ],
        "summary": "Find Buildable Decks",
        "description": "Find decks that can be built from a collection.",
        "operationId": "find_buildable_decks_recommendations_decks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FindDecksRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DeckSuggestion"
                  },
                  "type": "array",
                  "title": "Response Find Buildable Decks Recommendations Decks Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ComboSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "cards": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Cards"
          },
          "missing_cards": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Cards"
          },
          "produces": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Produces"
          },
          "bracket": {
            "type": "string",
            "title": "Bracket"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "completion_pct": {
            "type": "number",
            "title": "Completion Pct"
          }
        },
        "type": "object",
        "required": [
          "id",
          "cards",
          "missing_cards",
          "produces",
          "bracket",
          "score",
          "completion_pct"
        ],
        "title": "ComboSummary"
      },
...
      "DeckSuggestion": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "format": {
            "type": "string",
            "title": "Format"
          },
          "commander": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commander"
          },
          "archetype": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archetype"
          },
          "colors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Colors"
          },
          "key_cards_owned": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Key Cards Owned"
          },
          "key_cards_missing": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Key Cards Missing"
          },
          "completion_pct": {
            "type": "number",
            "title": "Completion Pct",
            "default": 0.0
          },
          "estimated_cost": {
            "type": "number",
            "title": "Estimated Cost",
            "default": 0.0
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reasons"
          },
          "near_combos": {
            "items": {
              "$ref": "#/components/schemas/ComboSummary"
            },
            "type": "array",
            "title": "Near Combos"
          },
          "complete_combos": {
            "items": {
              "$ref": "#/components/schemas/ComboSummary"
            },
            "type": "array",
            "title": "Complete Combos"
          },
          "combo_score": {
            "type": "number",
            "title": "Combo Score",
            "default": 0.0
          },
          "limited_bombs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Limited Bombs"
          },
          "filter_reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Filter Reasons"
          },
          "curve_warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Curve Warnings"
          },
          "interaction_count": {
            "type": "integer",
            "title": "Interaction Count",
            "default": 0
          },
          "quality_score": {
            "type": "number",
            "title": "Quality Score",
            "default": 0.0
          },
          "mana_base_quality": {
            "type": "string",
            "title": "Mana Base Quality",
            "default": ""
          },
          "fixing_land_count": {
            "type": "integer",
            "title": "Fixing Land Count",
            "default": 0
          },
          "win_condition_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Win Condition Types"
          },
          "tribal_strength": {
            "type": "string",
            "title": "Tribal Strength",
            "default": ""
          },
          "theme_strength": {
            "type": "string",
            "title": "Theme Strength",
            "default": ""
          },
          "lord_count": {
            "type": "integer",
            "title": "Lord Count",
            "default": 0
          },
          "ramp_count": {
            "type": "integer",
            "title": "Ramp Count",
            "default": 0
          },
          "ramp_warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Ramp Warnings"
          },
          "card_advantage_count": {
            "type": "integer",
            "title": "Card Advantage Count",
            "default": 0
          },
          "card_advantage_breakdown": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Card Advantage Breakdown"
          },
          "card_advantage_warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Card Advantage Warnings"
          }
        },
        "type": "object",
        "required": [
          "name",
          "format"
        ],
        "title": "DeckSuggestion"
      },
...
      "FindCommandersRequest": {
        "properties": {
          "collection_cards": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Collection Cards",
            "description": "Card names in collection (optional if use_collection=True)"
          },
          "use_collection": {
            "type": "boolean",
            "title": "Use Collection",
            "description": "Use entire user collection instead of passing card list",
            "default": false
          },
          "colors": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Colors",
            "description": "Filter by colors (W, U, B, R, G)"
          },
          "creature_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Creature Type",
            "description": "Filter by creature type"
          },
          "creature_types": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Creature Types",
            "description": "Filter by multiple creature types (tribals)"
          },
          "theme": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Theme",
            "description": "Filter by theme"
          },
          "themes": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Themes",
            "description": "Filter by multiple themes"
          },
          "format": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Format",
            "description": "Filter by format (commander, standard, modern, etc.)"
          },
          "set_codes": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Set Codes",
            "description": "Prioritize cards from these sets"
          },
          "limit": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Limit",
            "description": "Maximum commanders to return",
            "default": 10
          }
        },
        "type": "object",
        "title": "FindCommandersRequest",
        "description": "Request for finding commanders for a collection."
      },
      "FindDecksRequest": {
        "properties": {
          "collection_cards": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Collection Cards",
            "description": "Card names in collection (optional if use_collection=True)"
          },
          "use_collection": {
            "type": "boolean",
            "title": "Use Collection",
            "description": "Use entire user collection instead of passing card list",
            "default": false
          },
          "colors": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Colors",
            "description": "Filter by colors"
          },
          "creature_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Creature Type",
            "description": "Filter by tribal type"
          },
          "creature_types": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Creature Types",
            "description": "Filter by multiple creature types (tribals)"
          },
          "theme": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Theme",
            "description": "Filter by theme"
          },
          "themes": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Themes",
            "description": "Filter by multiple themes"
          },
          "format": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Format",
            "description": "Filter by format (commander, standard, modern, etc.)"
          },
          "set_codes": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Set Codes",
            "description": "Prioritize cards from these sets"
          },
          "min_completion": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Min Completion",
            "description": "Minimum deck completion percentage",
            "default": 0.0
          },
          "limit": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Limit",
            "description": "Maximum decks to return",
            "default": 10
          }
        },
        "type": "object",
        "title": "FindDecksRequest",
        "description": "Request for finding buildable decks."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SuggestCardsRequest": {
        "properties": {
          "deck_cards": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Deck Cards",
            "description": "Current deck card names"
          },
          "format_legal": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Format Legal",
            "description": "Format to filter by"
          },
          "budget_max": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Max",
            "description": "Maximum price per card"
          },
          "max_results": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Max Results",
            "description": "Maximum suggestions",
            "default": 10
          },
          "set_codes": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Set Codes",
            "description": "Prioritize cards from these sets"
          },
          "themes": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Themes",
            "description": "Filter by themes"
          },
          "creature_types": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Creature Types",
            "description": "Filter by creature types (tribals)"
          },
          "owned_only": {
            "type": "boolean",
            "title": "Owned Only",
            "description": "Only suggest cards from user's collection",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "deck_cards"
        ],
        "title": "SuggestCardsRequest",
        "description": "Request for card suggestions."
      },
      "SuggestCardsResult": {
        "properties": {
          "suggestions": {
            "items": {
              "$ref": "#/components/schemas/SuggestedCard"
            },
            "type": "array",
            "title": "Suggestions"
          },
          "detected_themes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Detected Themes"
          },
          "deck_colors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Deck Colors"
          }
        },
        "type": "object",
        "title": "SuggestCardsResult",
        "description": "Result of suggest_cards tool."
      },
      "SuggestedCard": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "category": {
            "type": "string",
            "enum": [
              "synergy",
              "staple",
              "upgrade",
              "budget"
            ],
            "title": "Category"
          },
          "mana_cost": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mana Cost"
          },
          "type_line": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type Line"
          },
          "price_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Usd"
          },
          "owned": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owned"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "tier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tier"
          },
          "edhrec_rank": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Edhrec Rank"
          },
          "combo_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Combo Count"
          }
        },
        "type": "object",
        "required": [
          "name",
          "reason",
          "category"
        ],
        "title": "SuggestedCard",
        "description": "A suggested card to add."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}